app/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
pydantic==2.5.2
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
firebase-admin==6.2.0
pytesseract==0.3.10
Pillow==10.1.0
pdfplumber==0.10.3
reportlab==4.0.7
aiofiles==23.2.1
